Cope with there not being a newline after the template (after '}}')
The regular expression that divides up the page was failing if there was
no whitespace after the template tag. This commit fixes that, so we make
no assumption at all about what follows the template tag. This means
updating some of the test, since there existing text which is considered
to be part of our output will have changed.

Fixes #1