Handle the (common) situation where nothing's after the terminator
It seems to be very common for there to be nothing at all (not even a
newline) after the terminating comment - it seems as if the trailing
whitespace in pages might be trimmed when you get it back from the
API (?). In any case, we currently aren't detecting the terminator in
those cases, which is bad, since you won't replace the content, but
instead insert it again. This commit fixes that.