Allow new content to contain formatting characters
If the new content previously contained, say, a percentage sign, this
would be interpolated into to the text to pass to the 'format' command,
and be taken as a formatting command, thus triggering an error.

Fix this by running everything through format, rather than a mix of
paramaters and interpolations.